+++ dfb/utils.py
@@ -18,8 +18,6 @@ logger = logging.getLogger(__name__)
 
 tsrep = namedtuple("timestamps", ("ts", "dt", "obj", "pretty"))
 
-_r = repr
-
 
 def time2all(dt_or_ts):
     """Convert from dt or ts to all formats"""