+++ tests/testutils.py
@@ -92,7 +92,7 @@ class Tester:
     def write_config(self):
         with open(self.configfile, "wt") as fobj:
             for key, val in self.config.items():
-                print(f"{key} = {_r(val)}", file=fobj)
+                print(f"{key} = {val!r}", file=fobj)
 
         self.config_obj = dfb.configuration.Config(self.configfile).parse()
         self.dstdb = dfb.dstdb.DFBDST(self.config_obj)